refactor: Refactor to separate view and logic of card component
