+++ src/components/cards/CardItem/index.jsx
@@ -1,6 +1,5 @@
 import Image from 'next/image';
 import classNames from 'classnames/bind';
-import Cards from '@/api/cards';
 import Avatar from '@/components/common/Avatar';
 import BaseButton from '@/components/common/button/BaseButton';
 import CreateCard from '@/components/dashboard/modal/card/CreateCard';
@@ -8,7 +7,8 @@ import CardTags from '@/components/common/CardTags';
 import CommonModal from '@/components/layout/modal/CommonModal';
 import DetailCard from '@/components/dashboard/modal/card/DetailCard';
 import IconModal from '@/components/layout/modal/IconModal';
-import useCardStore from '@/stores/useCardStore';
+import useCardFetchData from '@/hooks/logic/useCardFetchData';
+import useCreateCard from '@/hooks/logic/useCreateCard';
 import useModalState from '@/hooks/useModalState';
 import { IMAGE_REGEX } from '@/constants';
 import { ICON } from '@/constants/importImage';
@@ -18,38 +18,19 @@ const cx = classNames.bind(styles);
 const { calendar, remove } = ICON;
 
 const CardItem = ({ columnId, id, columnName }) => {
-  const cardList = useCardStore((prev) => prev.cardList[columnId]);
-  const cardItemData = cardList?.cards.find((card) => card.id === id);
-
+  const { cardItemData } = useCardFetchData({ columnId, id });
   const { assignee, title, imageUrl, tags, dueDate } = cardItemData;
-
   const { modalState, toggleModal } = useModalState([
     'detailCard',
     'editCard',
     'deleteCard',
   ]);
-
   const isImageUrl = !!imageUrl && IMAGE_REGEX.test(imageUrl.toLowerCase());
-
-  const handleModalOpen = (value) => {
-    if (!value) {
-      toggleModal('detailCard');
-      toggleModal('editCard');
-    } else {
-      toggleModal('deleteCard');
-    }
-  };
-
-  const handleDeleteModalClose = () => {
-    toggleModal('deleteCard');
-  };
-
-  const handleCardDelete = async () => {
-    await Cards.delete(id);
-    await Cards.getList(columnId);
-    toggleModal('deleteCard');
-    toggleModal('detailCard');
-  };
+  const { handleModalOpen, handleDeleteModalClose, handleCardDelete } = useCreateCard({
+    id,
+    columnId,
+    toggleModal,
+  });
 
   return (
     <>
